优化项目结构，优化构建流程，优化混淆流程，优化解析器内部类识别
