--- appAideBase/src/main/java/io/github/zeroaicy/util/SystemMemory.java
@@ -1,52 +1,0 @@
-package io.github.zeroaicy.util;
-import java.io.*;
-
-import android.app.ActivityManager;
-import android.content.Context;
-import android.text.format.Formatter;
-
-public class SystemMemory {
-    /**
-     *   * 获取android当前可用运行内存大小
-     *   * @param context
-     *   *
-     */
-    public static String getAvailMemory(Context context) {
-        ActivityManager am = (ActivityManager) context.getSystemService(Context.ACTIVITY_SERVICE);
-        ActivityManager.MemoryInfo mi = new ActivityManager.MemoryInfo();
-        am.getMemoryInfo(mi);
-// mi.availMem; 当前系统的可用内存
-        return Formatter.formatFileSize(context, mi.availMem);// 将获取的内存大小规格化
-    }
-
-
-
-
-    /**
-     *   * 获取android总运行内存大小
-     *   * @param context
-     *   *
-     */
-    public static String getTotalMemory(Context context) {
-        String str1 = "/proc/meminfo";// 系统内存信息文件
-        String str2;
-        String[] arrayOfString;
-        long initial_memory = 0;
-        try {
-            FileReader localFileReader = new FileReader(str1);
-            BufferedReader localBufferedReader = new BufferedReader(localFileReader, 8192);
-            str2 = localBufferedReader.readLine();// 读取meminfo第一行，系统总内存大小
-            arrayOfString = str2.split("\\s+");
-            for (String num : arrayOfString) {
-                Log.i(str2, num + "\t");
-            }
-            // 获得系统总内存，单位是KB
-            int i = Integer.valueOf(arrayOfString[1]).intValue();
-            //int值乘以1024转换为long类型
-            initial_memory = new Long((long)i*1024);
-            localBufferedReader.close();
-        } catch (IOException e) {
-        }
-        return Formatter.formatFileSize(context, initial_memory);// Byte转换为KB或者MB，内存大小规格化
-    }
-}
