--- appAideBase/src/main/java/io/github/zeroaicy/util/reflect/ReflectPieException.java
@@ -1,60 +1,0 @@
-package io.github.zeroaicy.util.reflect;
-
-/*
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-//package org.joor;
-
-import java.lang.reflect.InvocationTargetException;
-
-/**
- * A unchecked wrapper for any of Java's checked reflection exceptions:
- * <p>
- * These exceptions are
- * <ul>
- * <li> {@link ClassNotFoundException}</li>
- * <li> {@link IllegalAccessException}</li>
- * <li> {@link IllegalArgumentException}</li>
- * <li> {@link InstantiationException}</li>
- * <li> {@link InvocationTargetException}</li>
- * <li> {@link NoSuchMethodException}</li>
- * <li> {@link NoSuchFieldException}</li>
- * <li> {@link SecurityException}</li>
- * </ul>
- *
- * @author Lukas Eder
- */
-public class ReflectPieException extends RuntimeException {
-
-    /**
-     * Generated UID
-     */
-    private static final long serialVersionUID = -6213149635297151442L;
-
-    public ReflectPieException(String message) {
-        super(message);
-    }
-
-    public ReflectPieException(String message, Throwable cause) {
-        super(message, cause);
-    }
-
-    public ReflectPieException() {
-        super();
-    }
-
-    public ReflectPieException(Throwable cause) {
-        super(cause);
-    }
-}
-
