+++ app_rewrite/src/main/java/com/aide/ui/services/NativeCodeSupportService$q$a.java
@@ -10,13 +10,14 @@ import abcd.fy;
 import abcd.gy;
 import abcd.iy;
 import com.aide.ui.services.NativeCodeSupportService;
+import com.aide.ui.build.packagingservice.ExternalPackagingService;
 
 class NativeCodeSupportService$q$a implements Runnable {
     final boolean WB;
     final NativeCodeSupportService$q mb;
     public NativeCodeSupportService$q$a(NativeCodeSupportService$q qVar, boolean z) {
         this.mb = qVar;
-        this.WB = z;
+        this.WB = z;ExternalPackagingService m;
     }
 
     @Override
